[@mantine/core] TagInput: move to core
