+++ src/mantine-core/src/index.ts
@@ -81,6 +81,7 @@ export * from './Stepper';
 export * from './Switch';
 export * from './Table';
 export * from './Tabs';
+export * from './TagInput';
 export * from './Text';
 export * from './Textarea';
 export * from './TextInput';
