+++ src/mantine-demos/src/demos/core/TagInput/TagInput.demo.clearable.tsx
@@ -1,28 +1,29 @@
 import React from 'react';
 import { MantineDemo } from '@mantine/ds';
-import { TagInput } from '@mantine/labs';
+import { TagInput } from '@mantine/core';
 
 const code = `
 <TagInput
   label="Technologies that you're interested in"
-  placeholder="Add up to 5"
-  maxTags={5}
-/>
-`;
+  placeholder="Add as many as you like"
+  defaultValue={['react', 'next']}
+  clearable
+/>`;
 
 function Demo() {
   return (
     <div style={{ maxWidth: 400, marginLeft: 'auto', marginRight: 'auto' }}>
       <TagInput
         label="Technologies that you're interested in"
-        placeholder="Add up to 5"
-        maxTags={5}
+        placeholder="Add as many as you like"
+        defaultValue={['react', 'next']}
+        clearable
       />
     </div>
   );
 }
 
-export const maxTags: MantineDemo = {
+export const clearable: MantineDemo = {
   type: 'demo',
   code,
   component: Demo,
