+++ src/mantine-demos/src/demos/core/TagInput/TagInput.demo.maxTags.tsx
@@ -1,29 +1,28 @@
 import React from 'react';
 import { MantineDemo } from '@mantine/ds';
-import { TagInput } from '@mantine/labs';
+import { TagInput } from '@mantine/core';
 
 const code = `
 <TagInput
   label="Technologies that you're interested in"
-  placeholder="Add as many as you like"
-  defaultValue={['react', 'next']}
-  clearable
-/>`;
+  placeholder="Add up to 5"
+  maxTags={5}
+/>
+`;
 
 function Demo() {
   return (
     <div style={{ maxWidth: 400, marginLeft: 'auto', marginRight: 'auto' }}>
       <TagInput
         label="Technologies that you're interested in"
-        placeholder="Add as many as you like"
-        defaultValue={['react', 'next']}
-        clearable
+        placeholder="Add up to 5"
+        maxTags={5}
       />
     </div>
   );
 }
 
-export const clearable: MantineDemo = {
+export const maxTags: MantineDemo = {
   type: 'demo',
   code,
   component: Demo,
