+++ src/mantine-demos/src/demos/core/TagInput/TagInput.demo.usage.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
 import { MantineDemo } from '@mantine/ds';
-import { TagInput } from '@mantine/labs';
+import { TagInput } from '@mantine/core';
 
 const code = `
 <TagInput
   label="Technologies that you're interested in"
   placeholder="Add as many as you like"
-  defaultValue={['react', 'next']}
-  onlyUnique
 />
 `;
 
@@ -17,14 +15,12 @@ function Demo() {
       <TagInput
         label="Technologies that you're interested in"
         placeholder="Add as many as you like"
-        defaultValue={['react', 'next']}
-        onlyUnique
       />
     </div>
   );
 }
 
-export const onlyUnique: MantineDemo = {
+export const usage: MantineDemo = {
   type: 'demo',
   code,
   component: Demo,
