+++ src/mantine-labs/src/index.ts
@@ -1,2 +1,1 @@
 export * from './PinInput';
-export * from './TagInput';
